+++ controllers/katas/index.js
@@ -1,0 +1,9 @@
+function getAllKatas () {
+    console.log('getting all katas');
+}
+
+function getSingleKata () {
+    console.log('getting single katas');
+}
+
+module.exports = {getAllKatas, getSingleKata};
