+++ controllers/users/index.js
@@ -1,0 +1,9 @@
+function getAllUsers () {
+    console.log('getting all users');
+}
+
+function getSingleUser () {
+    console.log('getting single user');
+}
+
+module.exports = {getAllUsers, getSingleUser};
