+++ routes/kataRouter.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const {getAllKatas, getSingleKata} = require('../controllers');
+const {getAllKatas, getSingleKata} = require('../controllers/katas');
 
 router.get('/', getAllKatas);
 router.get('/:kata_name', getSingleKata);
